+++ src/apps/applib/lib/test.c
@@ -33,4 +33,17 @@ void t_assert(int value, const char* msg) {
     printf(".");
 }
 
+void t_assertBoundary(H3Index h3, const GeoBoundary* b1) {
+    // generate cell boundary for the h3 index
+    GeoBoundary b2;
+    H3_EXPORT(h3ToGeoBoundary)(h3, &b2);
+
+    t_assert(b1->numVerts == b2.numVerts, "expected cell boundary count");
+
+    for (int v = 0; v < b1->numVerts; v++) {
+        t_assert(geoAlmostEqual(&b1->verts[v], &b2.verts[v]),
+                 "got expected vertex");
+    }
+}
+
 int testCount() { return globalTestCount; }